+++ run-stash-backup-client.0.6.ps1
@@ -2,8 +2,8 @@
 #                                                                                                 #
 #  Script for Executing Stash Backup Client                                                       #
 #                                                                                                 #
-#  Version: 0.4                                                                                   #
-#  Date: 19.08.2014                                                                               #
+#  Version: 0.6                                                                                   #
+#  Date: 25.08.2014                                                                               #
 #  Author: Armin Pfäffle                                                                          #
 #  E-Mail mail@armin-pfaeffle.de                                                                  #
 #  Web: http://www.armin-pfaeffle.de                                                              #
@@ -18,7 +18,7 @@
 #
 # Logs a message to the log file and outputs it if $echo is $TRUE.
 #
-function Log($text, $withTimestamp = $TRUE, $echo = $TRUE)
+Function Log($text, $withTimestamp = $TRUE, $echo = $TRUE)
 {
 	$line = ""
 	if ($text) {
@@ -39,7 +39,7 @@ function Log($text, $withTimestamp = $TRUE, $echo = $TRUE)
 # Ensures that there is log directory given by $logDirectory. If folder does not
 # exist it is created.
 #
-function EnsureLogDirectory
+Function EnsureLogDirectory
 {
 	if( -not (Test-Path $logDirectory) )
 	{
@@ -51,7 +51,7 @@ function EnsureLogDirectory
 # Checks if mail credential file exists; if not asks user for credential and saves given
 # information in credential file. Password is encrypted.
 #
-function EnsureMailCredentialFile
+Function EnsureMailCredentialFile
 {
 	If (!(Test-Path $mailCredentialFilename))
 	{
@@ -83,7 +83,7 @@ function EnsureMailCredentialFile
 #
 # Returns date from mail credential file and returns PSCredential instance.
 #
-function GetMailCredential
+Function GetMailCredential
 {
 	Log "Load mail credential"
 
@@ -98,7 +98,7 @@ function GetMailCredential
 #
 # Sends mail with given subject. Adds additional text to body if $additionalBody is set.
 #
-function SendMail($subject, $body = "", $prependScriptStartAndEndTimestamp = $TRUE)
+Function SendMail($subject, $body = "", $prependScriptStartAndEndTimestamp = $TRUE)
 {
 	Log ("Sending mail '{0}'" -f $subject)
 
@@ -137,7 +137,7 @@ function SendMail($subject, $body = "", $prependScriptStartAndEndTimestamp = $TR
 # Waits for service to run. If service needs to long to start that FALSE is returned.
 # If service does not start at all, FALSE is returned.
 #
-function WaitForRunningService
+Function WaitForRunningService
 {
 	$startTimestamp = Get-Date
 	$service = Get-Service -Name $configuration["ServiceName"]
@@ -156,7 +156,7 @@ function WaitForRunningService
 # Changes working directory to Stash Backup Client directory, then runs the Stash Backup Client
 # and finally returns to the original working directory.
 #
-function RunStashBackup
+Function RunStashBackup
 {
 	Log "Run Stash Backup Client"
 	Log ""
@@ -176,6 +176,60 @@ function RunStashBackup
 	cd $currentDirectory
 
 	Return $output
+}
+
+#
+# Returns the file size in a human readable size. If file could not be extracted from mail output
+# or file does not exist, then $FALSE is returned.
+#
+Function GetBackupFileSize($backupOutput)
+{
+	$pattern = "Backup complete: (.*?)\.tar"
+	$regex = [regex] $pattern
+	$match = $regex.Match($backupOutput)
+	if ($match.Success -and ($match.Groups.Count -gt 1)) {
+		$filename = "{0}.tar" -f $match.Groups[1]
+		if (Test-Path $filename) {
+			$fileSize = (Get-Item $filename).length # in KB
+			Return FormatFileSize $fileSize
+		}
+	}
+	Return $FALSE
+}
+
+#
+# Formats file size so it is readable for humans.
+#
+Function FormatFileSize($size)
+{
+    If     ($size -gt 1TB) { Return "{0:0.00} TB" -f ($size / 1TB) }
+    ElseIf ($size -gt 1GB) { Return "{0:0.00} GB" -f ($size / 1GB) }
+    ElseIf ($size -gt 1MB) { Return "{0:0.00} MB" -f ($size / 1MB) }
+    ElseIf ($size -gt 1KB) { Return "{0:0.00} kB" -f ($size / 1KB) }
+    ElseIf ($size -gt 0)   { Return "{0:0.00} B" -f $size }
+    Else                   { Return "" }
+}
+
+#
+# Prepares the body for the final notification mail. Adds the computername and the backup
+# file size if backup was successful.
+#
+Function PrepareMailBody($successful, $backupOutput)
+{
+	if (!$configuration["Mail"]["AddLogToBody"]) {
+		$body = ""
+	} Else {
+		$computername = gc env:computername
+		$body = "Computername: {0}" -f $computername
+
+		if ($successful) {
+			$fileSize = GetBackupFileSize $backupOutput
+			$body = "`n{0}Size of backup: {1}" -f $body, $fileSize
+		}
+
+		$body = "{0}`n`n{1}" -f $body, $backupOutput
+	}
+	Return $body
 }
 
 
@@ -212,18 +266,18 @@ if ($service)
 		Log "Service is running"
 		Log "Starting Stash Backup Client"
 
-		$mailOutput = RunStashBackup
+		$backupOutput = RunStashBackup
 
 		if ($configuration["SendMailAfterBackup"]) {
-			if (!$configuration["Mail"]["AddLogToBody"]) {
-				$mailOutput = ""
-			}
-			if ($mailOutput -match "Exception") {
+			if ($backupOutput -match "Exception") {
 				$subject = $configuration["Mail"]["Subject"]["Error"]
+				$successful = $FALSE
 			} Else {
 				$subject = $configuration["Mail"]["Subject"]["Success"]
+				$successful = $TRUE
 			}
-			$sendResult = SendMail $subject $mailOutput
+			$body = PrepareMailBody $successful $backupOutput
+			$sendResult = SendMail $subject $body
 		}
 	} Else {
 		Log "Service is not running, needs to long to start or is stopped and does not start"