+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 Changelog
 ==========================
 
+### Version 0.6
+
+* Added computer name to final notification mail
+* Added backup file size to final notification mail
+
+
 ### Version 0.5
 
 * Added option: send mail before backup starts
 * Improved SendMail function
+
 
 ### Version 0.4
 
 * After entering mail credential a test mail is sent
 * If test mail fails, credential file is deleted
 * SendMail now returns $TRUE or $FALSE
+
 
 ### Version < 0.4